The code changes primarily involve the integration of SignalR into the application.
Also, modification of CORS policy, and changes to the methods in the `IService` interface and `Service` class. The `ItemHub` class was removed, and a new `ItemsHub` class was created. The `Service` and `ServiceDecorator` classes were registered as scoped services. The `DeleteItem` and `CreateItem` methods were changed to be asynchronous. The `ItemsHub` was mapped to the `/itemsHub` route. The `CreateItem`, `DeleteItem`, and `GetItems` methods in the `IService` interface and `Service` class were changed to return `ValueTask`. The `SignalRTemplate.Infrastructure` project was added as a reference and to the solution file. The `package-lock.json` and `package.json` files were updated with new dependencies. The `App.tsx` file was updated to create and start a new `HubConnection` instance. The `jsx` option in the `tsconfig.json` file was changed from `react-jsx` to `preserve`. A new `ServiceDecorator.cs` file was created that defines a `ServiceDecorator` class.

Changes:

1. Removal of `ItemHub` class from `ItemHub.cs` file.
2. Addition of `Microsoft.AspNetCore.SignalR` and `SignalRTemplate.Infrastructure` namespaces to `Program.cs` file.
3. Modification of CORS policy in `Program.cs` file.
4. Registration of `Service` and `ServiceDecorator` classes as scoped services in `Program.cs` file.
5. Addition of SignalR service to `Program.cs` file.
6. Changes to `DeleteItem` and `CreateItem` methods in `Program.cs` file to be asynchronous.
7. Mapping of `ItemsHub` to `/itemsHub` route in `Program.cs` file.
8. Changes to `CreateItem`, `DeleteItem`, and `GetItems` methods in `IService` interface and `Service` class to return `ValueTask`.
9. Addition of `SignalRTemplate.Infrastructure` project as a reference in `SignalRTemplate.WebApi.csproj` file and to `SignalRTemplate.sln` file.
10. Updates to `package-lock.json` and `package.json` files with new dependencies.
11. Updates to `App.tsx` file to create and start a new `HubConnection` instance.
12. Changes to `jsx` option in `tsconfig.json` file from `react-jsx` to `preserve`.
13. Creation of new `ItemsHub.cs` and `ServiceDecorator.cs` files.